Fix build warnings in x64.
A std::min/max with size_t and unsigned int.

Popped up in f98e63a54b8a2.

BUG=

Reviewed-on: https://chromium-review.googlesource.com/200484
Tested-by: Jamie Madill <jmadill@chromium.org>

+++ src/libGLESv2/renderer/d3d11/InputLayoutCache.cpp
@@ -207,8 +207,8 @@ GLenum InputLayoutCache::applyVertexBuffers(TranslatedAttribute attributes[gl::M
             vertexOffset != mCurrentVertexOffsets[i])
         {
             dirtyBuffers = true;
-            minDiff = std::min(minDiff, i);
-            maxDiff = std::max(maxDiff, i);
+            minDiff = std::min(minDiff, static_cast<size_t>(i));
+            maxDiff = std::max(maxDiff, static_cast<size_t>(i));
 
             mCurrentBuffers[i] = buffer;
             mCurrentVertexStrides[i] = vertexStride;
